Refactor to new renderer done
