+++ src/Directory.Build.props
@@ -23,9 +23,9 @@
 		<Deterministic>true</Deterministic>
 	</PropertyGroup>
 
-	<PropertyGroup Condition="'$(Configuration)' == 'Release' AND $(MSBuildProjectName)!='bunit.template' AND $(MSBuildProjectName)!='bunit'">
+	<!--<PropertyGroup Condition="'$(Configuration)' == 'Release' AND $(MSBuildProjectName)!='bunit.template' AND $(MSBuildProjectName)!='bunit'">
 		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-	</PropertyGroup>
+	</PropertyGroup>-->
 
 	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1' AND $(MSBuildProjectName)!='bunit.template' AND $(MSBuildProjectName)!='bunit'">
 		<PackageReference Include="Microsoft.AspNetCore.Components.Authorization" Version="3.1.1" />
