+++ src/bunit.core/Rendering/IRenderedComponentActivator.cs
@@ -1,0 +1,16 @@
+
+using Microsoft.AspNetCore.Components;
+
+namespace Bunit.Rendering
+{
+	public interface IRenderedComponentActivator
+	{
+		IRenderedFragmentBase CreateRenderedComponent(int componentId);
+
+		IRenderedComponentBase<TComponent> CreateRenderedComponent<TComponent>(int componentId)
+			where TComponent : IComponent;
+
+		IRenderedComponentBase<TComponent> CreateRenderedComponent<TComponent>(int componentId, TComponent component, RenderTreeFrameCollection componentFrames)
+			where TComponent : IComponent;
+	}
+}
