+++ src/bunit.core/Rendering/TestRenderer.cs
@@ -1,166 +1,263 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Runtime.ExceptionServices;
+using System.Text;
 using System.Threading.Tasks;
-
-using Bunit.Extensions;
 
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.RenderTree;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
 namespace Bunit.Rendering
 {
-	/// <summary>
-	/// Generalized Blazor renderer for testing purposes.
-	/// </summary>
-	public class TestRenderer : Renderer, ITestRenderer, IRenderEventProducer
+	public sealed class TestRenderer : Renderer, ITestRenderer
 	{
+		private readonly object _renderTreeAccessLock = new object();
 		private readonly ILogger _logger;
-		private readonly List<IRenderEventHandler> _renderEventHandlers = new List<IRenderEventHandler>();
+		private readonly IRenderedComponentActivator _activator;
 		private Exception? _unhandledException;
+		private readonly Dictionary<int, IRenderedFragmentBase> _renderedComponents = new Dictionary<int, IRenderedFragmentBase>();
 
-		/// <inheritdoc/>
 		public override Dispatcher Dispatcher { get; } = Dispatcher.CreateDefault();
 
-		/// <summary>
-		/// Creates an instance of the <see cref="TestRenderer"/> class.
-		/// </summary>
-		public TestRenderer(IServiceProvider services, ILoggerFactory loggerFactory) : base(services, loggerFactory)
+		public TestRenderer(IRenderedComponentActivator activator, IServiceProvider services, ILoggerFactory loggerFactory) : base(services, loggerFactory)
 		{
 			_logger = loggerFactory.CreateLogger<TestRenderer>();
+			_activator = activator;
 		}
 
-		/// <summary>
-		/// Adds a <see cref="IRenderEventHandler"/> to this renderer,
-		/// which will be triggered when the renderer has finished rendering
-		/// a render cycle.
-		/// </summary>
-		/// <param name="handler">The handler to add.</param>
-		public void AddRenderEventHandler(IRenderEventHandler handler) => _renderEventHandlers.Add(handler);
-
-		/// <summary>
-		/// Removes a <see cref="IRenderEventHandler"/> from this renderer.
-		/// </summary>
-		/// <param name="handler">The handler to remove.</param>
-		public void RemoveRenderEventHandler(IRenderEventHandler handler) => _renderEventHandlers.Remove(handler);
-
-		/// <inheritdoc/>
-		public (int ComponentId, TComponent Component) RenderComponent<TComponent>(IEnumerable<ComponentParameter> parameters) where TComponent : IComponent
+		public IRenderedFragmentBase RenderFragment(RenderFragment renderFragment)
 		{
-			var componentType = typeof(TComponent);
-			var renderFragment = parameters.ToComponentRenderFragment<TComponent>();
-			var wrapperId = RenderFragmentInsideWrapper(renderFragment);
-			return FindComponent<TComponent>(wrapperId);
+			IRenderedFragmentBase renderedComponent = default!;
+			var task = Dispatcher.InvokeAsync(() =>
+			{
+				var root = new WrapperComponent(renderFragment);
+				var rootComponentId = AssignRootComponentId(root);
+				renderedComponent = _activator.CreateRenderedComponent(rootComponentId);
+				_renderedComponents.Add(rootComponentId, renderedComponent);
+				root.Render();
+			});
+
+			Debug.Assert(task.IsCompleted, "The render task did not complete as expected");
+			AssertNoUnhandledExceptions();
+
+			return renderedComponent!;
 		}
 
-		/// <inheritdoc/>
-		public int RenderFragment(RenderFragment renderFragment)
+		public IRenderedComponentBase<TComponent> RenderComponent<TComponent>(IEnumerable<ComponentParameter> componentParameters)
+			where TComponent : IComponent
 		{
-			return RenderFragmentInsideWrapper(renderFragment);
+			IRenderedComponentBase<TComponent> renderedComponent = default!;
+			var task = Dispatcher.InvokeAsync(() =>
+			{
+				var root = new WrapperComponent(componentParameters.ToComponentRenderFragment<TComponent>());
+				var rootComponentId = AssignRootComponentId(root);
+				renderedComponent = _activator.CreateRenderedComponent<TComponent>(rootComponentId);
+				_renderedComponents.Add(rootComponentId, renderedComponent);
+				root.Render();
+			});
+
+			Debug.Assert(task.IsCompleted, "The render task did not complete as expected");
+			AssertNoUnhandledExceptions();
+
+			return renderedComponent!;
 		}
 
-		/// <inheritdoc/>
-		public (int ComponentId, TComponent Component) FindComponent<TComponent>(int parentComponentId)
+		public IRenderedComponentBase<TComponent> FindComponent<TComponent>(IRenderedFragmentBase parentComponent) where TComponent : IComponent
 		{
-			var result = GetComponent<TComponent>(parentComponentId);
-			if (result.HasValue)
-				return result.Value;
-			else
-				throw new ComponentNotFoundException(typeof(TComponent));
+			if (parentComponent is null)
+				throw new ArgumentNullException(nameof(parentComponent));
+
+			var framesCollection = new RenderTreeFrameCollection();
+
+			lock (_renderTreeAccessLock)
+			{
+				if (TryFindComponent(parentComponent.ComponentId, out var id, out var component))
+				{
+					LoadRenderTreeFrames(id, framesCollection);
+					var rc = _activator.CreateRenderedComponent(id, component, framesCollection);
+					_renderedComponents.Add(rc.ComponentId, rc);
+					return rc;
+				}
+				else
+				{
+					throw new ComponentNotFoundException(typeof(TComponent));
+				}
+			}
+
+			bool TryFindComponent(int parentComponentId, out int componentId, out TComponent component)
+			{
+				var result = false;
+				componentId = -1;
+				component = default!;
+
+				var frames = LoadAndGetRenderTreeFrame(framesCollection, parentComponentId);
+
+				for (var i = 0; i < frames.Count; i++)
+				{
+					ref var frame = ref frames.Array[i];
+					if (frame.FrameType == RenderTreeFrameType.Component)
+					{
+						if (frame.Component is TComponent c)
+						{
+							componentId = frame.ComponentId;
+							component = c;
+							result = true;
+							break;
+						}
+
+						if (TryFindComponent(frame.ComponentId, out componentId, out component))
+						{
+							result = true;
+							break;
+						}
+					}
+				}
+
+				return result;
+			}
 		}
 
-		/// <inheritdoc/>
-		public IReadOnlyList<(int ComponentId, TComponent Component)> FindComponents<TComponent>(int parentComponentId)
+		public IReadOnlyList<IRenderedComponentBase<TComponent>> FindComponents<TComponent>(IRenderedFragmentBase parentComponent)
+			where TComponent : IComponent
 		{
-			return GetComponents<TComponent>(parentComponentId);
+			if (parentComponent is null)
+				throw new ArgumentNullException(nameof(parentComponent));
+
+			var result = new List<IRenderedComponentBase<TComponent>>();
+			var framesCollection = new RenderTreeFrameCollection();
+
+			lock (_renderTreeAccessLock)
+			{
+				FindComponentsInternal(parentComponent.ComponentId);
+				foreach (var rc in result)
+				{
+					_renderedComponents.Add(rc.ComponentId, rc);
+				}
+			}
+
+			return result;
+
+			void FindComponentsInternal(int componentId)
+			{
+				var frames = LoadAndGetRenderTreeFrame(framesCollection, componentId);
+
+				for (var i = 0; i < frames.Count; i++)
+				{
+					ref var frame = ref frames.Array[i];
+					if (frame.FrameType == RenderTreeFrameType.Component)
+					{
+						if (frame.Component is TComponent component)
+						{
+							var id = frame.ComponentId;
+							LoadRenderTreeFrames(id, framesCollection);
+							var rc = _activator.CreateRenderedComponent(id, component, framesCollection);
+							result.Add(rc);
+						}
+
+						FindComponentsInternal(frame.ComponentId);
+					}
+				}
+			}
 		}
 
-		/// <inheritdoc/>
-		public new ArrayRange<RenderTreeFrame> GetCurrentRenderTreeFrames(int componentId)
-		{
-			return base.GetCurrentRenderTreeFrames(componentId);
-		}
-
-		/// <inheritdoc/>
 		public new Task DispatchEventAsync(ulong eventHandlerId, EventFieldInfo fieldInfo, EventArgs eventArgs)
 		{
 			if (fieldInfo is null)
 				throw new ArgumentNullException(nameof(fieldInfo));
 
-			_logger.LogDebug(new EventId(10, nameof(DispatchEventAsync)), $"Starting trigger of '{fieldInfo.FieldValue}'");
-
 			var result = Dispatcher.InvokeAsync(() => base.DispatchEventAsync(eventHandlerId, fieldInfo, eventArgs));
 
 			AssertNoUnhandledExceptions();
 
-			if (result.IsCompletedSuccessfully)
-			{
-				_logger.LogDebug(new EventId(11, nameof(DispatchEventAsync)), $"Finished trigger synchronously for '{fieldInfo.FieldValue}'");
-			}
-			else
-			{
-				_logger.LogDebug(new EventId(13, nameof(DispatchEventAsync)), $"Event handler for '{fieldInfo.FieldValue}' returned an incomplete task with status {result.Status}");
-				result = result.ContinueWith(x =>
-				{
-					if (x.IsCompletedSuccessfully)
-					{
-						_logger.LogDebug(new EventId(12, nameof(DispatchEventAsync)), $"Finished trigger asynchronously for '{fieldInfo.FieldValue}'");
-					}
-				}, TaskScheduler.Default);
-			}
-
 			return result;
 		}
 
-		private int RenderFragmentInsideWrapper(RenderFragment renderFragment)
+		/// <inheritdoc/>
+		protected override void ProcessPendingRender()
 		{
-			var wrapper = new WrapperComponent(renderFragment);
-
-			var wrapperId = AssignRootComponentId(wrapper);
-			AssertNoUnhandledExceptions();
-
-			Dispatcher.InvokeAsync(wrapper.Render).Wait();
-			AssertNoUnhandledExceptions();
-
-			return wrapperId;
+			lock (_renderTreeAccessLock)
+			{
+				base.ProcessPendingRender();
+			}
 		}
 
 		/// <inheritdoc/>
-		protected override void HandleException(Exception exception) => _unhandledException = exception;
+		protected override void HandleException(Exception exception)
+			=> _unhandledException = exception;
 
 		/// <inheritdoc/>
 		protected override Task UpdateDisplayAsync(in RenderBatch renderBatch)
 		{
-			_logger.LogDebug(new EventId(0, nameof(UpdateDisplayAsync)), $"New render batch with ReferenceFrames = {renderBatch.ReferenceFrames.Count}, UpdatedComponents = {renderBatch.UpdatedComponents.Count}, DisposedComponentIDs = {renderBatch.DisposedComponentIDs.Count}, DisposedEventHandlerIDs = {renderBatch.DisposedEventHandlerIDs.Count}");
+			var renderEvent = new RenderEvent(renderBatch, new RenderTreeFrameCollection());
 
-			return _renderEventHandlers.Count == 0
-				? Task.CompletedTask
-				: PublishRenderEvent(in renderBatch);
-		}
-
-		private Task PublishRenderEvent(in RenderBatch renderBatch)
-		{
-			var renderEvent = new RenderEvent(in renderBatch, this);
-
-			return _renderEventHandlers.Count switch
+			// removes disposed components
+			for (var i = 0; i < renderBatch.DisposedComponentIDs.Count; i++)
 			{
-				0 => Task.CompletedTask,
-				1 => _renderEventHandlers[0].Handle(renderEvent),
-				_ => NotifyEventHandlers(renderEvent)
-			};
+				var id = renderBatch.DisposedComponentIDs.Array[i];
+				if (_renderedComponents.TryGetValue(id, out var rc))
+				{
+					_renderedComponents.Remove(id);
+					rc.OnRender(renderEvent);
+				}
+			}
+
+			foreach (var (key, rc) in _renderedComponents.ToArray())
+			{
+				LoadRenderTreeFrames(rc.ComponentId, renderEvent.Frames);
+				rc.OnRender(renderEvent);
+
+				// RC can replace the instance of the component is bound
+				// to while processing the update event. 
+				if (key != rc.ComponentId)
+				{
+					_renderedComponents.Remove(key);
+					_renderedComponents.Add(rc.ComponentId, rc);
+				}
+			}
+
+			return Task.CompletedTask;
 		}
 
-		private Task NotifyEventHandlers(RenderEvent renderEvent)
+		/// <inheritdoc/>
+		protected override void Dispose(bool disposing)
 		{
-			// copy to new array since _renderEventHandlers might be modified by the
-			// Handle method on event handlers if component is disposed.
-			var handleTasks = _renderEventHandlers
-				.ToArray()
-				.Select(x => x.Handle(renderEvent));
+			if (disposing)
+			{
+				foreach (var rc in _renderedComponents.Values)
+				{
+					rc.Dispose();
+				}
+				_renderedComponents.Clear();
+			}
+			base.Dispose(disposing);
+		}
 
-			return Task.WhenAll(handleTasks);
+		private void LoadRenderTreeFrames(int componentId, RenderTreeFrameCollection framesCollection)
+		{
+			var frames = LoadAndGetRenderTreeFrame(framesCollection, componentId);
+
+			for (var i = 0; i < frames.Count; i++)
+			{
+				ref var frame = ref frames.Array[i];
+				if (frame.FrameType == RenderTreeFrameType.Component)
+				{
+					LoadRenderTreeFrames(frame.ComponentId, framesCollection);
+				}
+			}
+		}
+
+		private ArrayRange<RenderTreeFrame> LoadAndGetRenderTreeFrame(RenderTreeFrameCollection framesCollection, int componentId)
+		{
+			if (!framesCollection.Contains(componentId))
+			{
+				framesCollection.Add(componentId, GetCurrentRenderTreeFrames(componentId));
+			}
+
+			return framesCollection[componentId];
 		}
 
 		private void AssertNoUnhandledExceptions()
@@ -171,52 +268,6 @@ namespace Bunit.Rendering
 				var evt = new EventId(3, nameof(AssertNoUnhandledExceptions));
 				_logger.LogError(evt, unhandled, $"An unhandled exception happened during rendering: {unhandled.Message}{Environment.NewLine}{unhandled.StackTrace}");
 				ExceptionDispatchInfo.Capture(unhandled).Throw();
-			}
-		}
-
-		private (int ComponentId, TComponent Component)? GetComponent<TComponent>(int rootComponentId)
-		{
-			var ownFrames = GetCurrentRenderTreeFrames(rootComponentId);
-
-			for (var i = 0; i < ownFrames.Count; i++)
-			{
-				ref var frame = ref ownFrames.Array[i];
-				if (frame.FrameType == RenderTreeFrameType.Component)
-				{
-					if (frame.Component is TComponent component)
-						return (frame.ComponentId, component);
-
-					var result = GetComponent<TComponent>(frame.ComponentId);
-					if (result is { })
-						return result;
-				}
-			}
-
-			return null;
-		}
-
-		private IReadOnlyList<(int ComponentId, TComponent Component)> GetComponents<TComponent>(int rootComponentId)
-		{
-			var result = new List<(int ComponentId, TComponent Component)>();
-
-			GetComponentsInternal(rootComponentId, result);
-
-			return result;
-
-			void GetComponentsInternal(int rootComponentId, List<(int ComponentId, TComponent Component)> result)
-			{
-				var ownFrames = GetCurrentRenderTreeFrames(rootComponentId);
-				for (var i = 0; i < ownFrames.Count; i++)
-				{
-					ref var frame = ref ownFrames.Array[i];
-					if (frame.FrameType == RenderTreeFrameType.Component)
-					{
-						if (frame.Component is TComponent component)
-							result.Add((frame.ComponentId, component));
-
-						GetComponentsInternal(frame.ComponentId, result);
-					}
-				}
 			}
 		}
 	}
