+++ src/bunit.web/Extensions/TestServiceProviderExtensions.cs
@@ -28,6 +28,7 @@ namespace Bunit.Extensions
 					srv.GetRequiredService<HtmlComparer>()
 				)
 			);
+			services.AddSingleton<IRenderedComponentActivator, RenderedComponentActivator>();
 			return services;
 		}
 	}
