+++ src/bunit.web/RazorTesting/SnapshotTest.cs
@@ -53,11 +53,11 @@ namespace Bunit
 			if (SetupAsync is { })
 				await TryRunAsync(SetupAsync, this).ConfigureAwait(false);
 
-			var testRenderId = Renderer.RenderFragment(TestInput!);
-			var inputHtml = Htmlizer.GetHtml(Renderer, testRenderId);
+			var renderedTestInput = (IRenderedFragment)Renderer.RenderFragment(TestInput!);
+			var inputHtml = renderedTestInput.Markup;
 
-			var expectedRenderId = Renderer.RenderFragment(ExpectedOutput!);
-			var expectedHtml = Htmlizer.GetHtml(Renderer, expectedRenderId);
+			var renderedExpectedRender = (IRenderedFragment)Renderer.RenderFragment(ExpectedOutput!);
+			var expectedHtml = renderedExpectedRender.Markup;
 
 			VerifySnapshot(inputHtml, expectedHtml);
 		}
